Move validations to validation.js
