+++ backend/routes/api/session.js
@@ -1,22 +1,11 @@
 const express = require('express');
 const { check } = require('express-validator');
 const { setTokenCookie, restoreUser } = require('../../utils/auth');
-const { handleValidationErrors } = require('../../utils/validation');
+const { handleValidationErrors, validateLogin } = require('../../utils/validation');
 const { User } = require('../../db/models');
 
 const router = express.Router();
 
-// validate logins
-const validateLogin = [
-  check('credential')
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage('Please provide a valid email or username.'),
-  check('password')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a password.'),
-  handleValidationErrors
-];
 
 // Restore session user
 router.get('/', restoreUser, (req, res) => {
